drone_fault_detection_model/requirements.txt
# --- PyTorch & Vision ---
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
timm>=0.9.12

# --- Data Processing & Imaging ---
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# --- Audio Processing (Spectrogram Generation) ---
librosa>=0.10.0
soundfile>=0.12.0

# --- Training & Metrics ---
scikit-learn>=1.3.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# --- API Server ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# --- Model Export ---
onnx>=1.14.0
onnxruntime>=1.16.0
pyyaml>=6.0
